+++ spec/backbone.dualstorage.js
@@ -323,7 +323,7 @@ onlineSync = function(method, model, options) {
 };
 
 dualsync = function(method, model, options) {
-  var error, local, relayErrorCallback, success, temporaryId;
+  var error, hasOfflineStatusCode, local, relayErrorCallback, success, temporaryId, useOfflineStorage;
   options.storeName = getStoreName(model.collection, model);
   options.storeExists = Store.exists(options.storeName);
   options.success = callbackTranslator.forDualstorageCaller(options.success, model, options);
@@ -339,28 +339,37 @@ dualsync = function(method, model, options) {
   options.ignoreCallbacks = true;
   success = options.success;
   error = options.error;
-  relayErrorCallback = function(response) {
-    var offline, offlineStatusCodes, _ref;
+  useOfflineStorage = function() {
+    options.dirty = true;
+    return success(localsync(method, model, options));
+  };
+  hasOfflineStatusCode = function(xhr) {
+    var offlineStatusCodes, _ref;
     offlineStatusCodes = Backbone.DualStorage.offlineStatusCodes;
     if (_.isFunction(offlineStatusCodes)) {
-      offlineStatusCodes = offlineStatusCodes(response);
+      offlineStatusCodes = offlineStatusCodes(xhr);
     }
-    offline = response.status === 0 || (_ref = response.status, __indexOf.call(offlineStatusCodes, _ref) >= 0);
-    if (!offline || method === 'read' && !options.storeExists) {
-      return error(response);
+    return xhr.status === 0 || (_ref = xhr.status, __indexOf.call(offlineStatusCodes, _ref) >= 0);
+  };
+  relayErrorCallback = function(xhr) {
+    var online;
+    online = !hasOfflineStatusCode(xhr);
+    if (online || method === 'read' && !options.storeExists) {
+      return error(xhr);
     } else {
-      options.dirty = true;
-      return success(localsync(method, model, options));
+      return useOfflineStorage();
     }
   };
   switch (method) {
     case 'read':
       if (localsync('hasDirtyOrDestroyed', model, options)) {
-        options.dirty = true;
-        return success(localsync(method, model, options));
+        return useOfflineStorage();
       } else {
         options.success = function(resp, status, xhr) {
           var collection, idAttribute, modelAttributes, responseModel, _i, _len;
+          if (hasOfflineStatusCode(xhr)) {
+            return useOfflineStorage();
+          }
           resp = parseRemoteResponse(model, resp);
           if (model instanceof Backbone.Collection) {
             collection = model;
@@ -384,8 +393,8 @@ dualsync = function(method, model, options) {
           }
           return success(resp, status, xhr);
         };
-        options.error = function(resp) {
-          return relayErrorCallback(resp);
+        options.error = function(xhr) {
+          return relayErrorCallback(xhr);
         };
         return onlineSync(method, model, options);
       }
@@ -393,12 +402,15 @@ dualsync = function(method, model, options) {
     case 'create':
       options.success = function(resp, status, xhr) {
         var updatedModel;
+        if (hasOfflineStatusCode(xhr)) {
+          return useOfflineStorage();
+        }
         updatedModel = modelUpdatedWithResponse(model, resp);
         localsync(method, updatedModel, options);
         return success(resp, status, xhr);
       };
-      options.error = function(resp) {
-        return relayErrorCallback(resp);
+      options.error = function(xhr) {
+        return relayErrorCallback(xhr);
       };
       return onlineSync(method, model, options);
     case 'update':
@@ -406,6 +418,9 @@ dualsync = function(method, model, options) {
         temporaryId = model.id;
         options.success = function(resp, status, xhr) {
           var updatedModel;
+          if (hasOfflineStatusCode(xhr)) {
+            return useOfflineStorage();
+          }
           updatedModel = modelUpdatedWithResponse(model, resp);
           model.set(model.idAttribute, temporaryId, {
             silent: true
@@ -414,11 +429,11 @@ dualsync = function(method, model, options) {
           localsync('create', updatedModel, options);
           return success(resp, status, xhr);
         };
-        options.error = function(resp) {
+        options.error = function(xhr) {
           model.set(model.idAttribute, temporaryId, {
             silent: true
           });
-          return relayErrorCallback(resp);
+          return relayErrorCallback(xhr);
         };
         model.set(model.idAttribute, null, {
           silent: true
@@ -427,12 +442,15 @@ dualsync = function(method, model, options) {
       } else {
         options.success = function(resp, status, xhr) {
           var updatedModel;
+          if (hasOfflineStatusCode(xhr)) {
+            return useOfflineStorage();
+          }
           updatedModel = modelUpdatedWithResponse(model, resp);
           localsync(method, updatedModel, options);
           return success(resp, status, xhr);
         };
-        options.error = function(resp) {
-          return relayErrorCallback(resp);
+        options.error = function(xhr) {
+          return relayErrorCallback(xhr);
         };
         return onlineSync(method, model, options);
       }
@@ -443,11 +461,14 @@ dualsync = function(method, model, options) {
         return localsync(method, model, options);
       } else {
         options.success = function(resp, status, xhr) {
+          if (hasOfflineStatusCode(xhr)) {
+            return useOfflineStorage();
+          }
           localsync(method, model, options);
           return success(resp, status, xhr);
         };
-        options.error = function(resp) {
-          return relayErrorCallback(resp);
+        options.error = function(xhr) {
+          return relayErrorCallback(xhr);
         };
         return onlineSync(method, model, options);
       }